requirements.txt
flask>=3.0.0
anthropic>=0.40.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
pytz>=2024.1
requests>=2.31.0
gunicorn>=21.2.0
